requirements.txt
opencv-python
matplotlib
numpy
PyYAML